# server/public/swagger.json
{
	"components": {
...
		"schemas": {
...
			"QueueInterface": {
				"properties": {
					"userInQueueList": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"guid": {
						"type": "string"
					},
					"inHandUserId": {
						"type": "string",
						"nullable": true,
						"description": "В руках пользователя по guid"
					},
					"lastChangeUtc": {
						"type": "string",
						"format": "date-time",
						"description": "Время последнего изменения - взятие в руки или сдача устройства"
					}
				},
				"required": [
					"userInQueueList",
					"guid",
					"inHandUserId",
					"lastChangeUtc"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "access_token",
				"in": "header",
				"flow": "implicit",
				"scopes": {}
			}
		}
	},
	"info": {
		"title": "server",
		"version": "0.0.0",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/v1/queue/join": {
			"post": {
				"operationId": "StandInQueue",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/QueueInterface"
								}
							}
						}
					}
				},
				"tags": [
					"Queue"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"deviceGuid": {
										"type": "string"
									}
								},
								"required": [
									"deviceGuid"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/v1/queue/take": {
			"post": {
				"operationId": "TakeDevice",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/QueueInterface"
								}
							}
						}
					}
				},
				"tags": [
					"Queue"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"deviceGuid": {
										"type": "string"
									}
								},
								"required": [
									"deviceGuid"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/v1/queue/takeoff": {
			"post": {
				"operationId": "LaydownDevice",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/QueueInterface"
								}
							}
						}
					}
				},
				"tags": [
					"Queue"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"deviceGuid": {
										"type": "string"
									}
								},
								"required": [
									"deviceGuid"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/v1/queue/cancel": {
			"post": {
				"operationId": "CancelQueue",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/QueueInterface"
								}
							}
						}
					}
				},
				"tags": [
					"Queue"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"deviceGuid": {
										"type": "string"
									}
								},
								"required": [
									"deviceGuid"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/v1/queue/meta": {
			"get": {
				"operationId": "GetQueryInfo",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/QueueInterface"
								}
							}
						}
					}
				},
				"tags": [
					"Queue"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "deviceGuid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/queue/myqueue": {
			"get": {
				"operationId": "GetAllUserQueue",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/QueueInterface"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Queue"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/api/v1/queue/drop": {
			"post": {
				"operationId": "DropQuery",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Queue"
				],
				"security": [
					{
						"jwt": [
							"root",
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"deviceGuid": {
										"type": "string"
									}
								},
								"required": [
									"deviceGuid"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	],
...
}